requirements.txt
opencv-python>=4.8.0
tensorflow>=2.16.0
numpy>=1.24.0
mediapipe>=0.10.0
pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
imutils>=0.5.4